# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/moderation/queue": {
      "get": {
        "operationId": "ModerationController_getQueue",
        "summary": "Get moderation queue",
        "description": "Retrieve all pending items waiting for community review. Any logged user can view and review these items.",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "objectType",
            "required": false,
            "in": "query",
            "description": "Filter by object type (molecule, protein, reaction, article)",
            "schema": {
              "enum": [
                "molecule",
                "protein",
                "reaction",
                "article"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Queue retrieved successfully"
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/moderation/versions/{versionId}/review": {
      "post": {
        "operationId": "ModerationController_submitReview",
        "summary": "Submit community review",
        "description": "Submit your review (approve/reject) for a pending version. Your reputation score will be added to the total. Versions auto-approve when approval votes reach 100 reputation points, or auto-reject when rejection votes reach 100.",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "description": "Version ID to review",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVersionReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review submitted successfully"
          },
          "403": {
            "description": "Cannot review own changes or already reviewed"
          },
          "404": {
            "description": "Version not found"
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/moderation/versions/{versionId}/reviews": {
      "get": {
        "operationId": "ModerationController_getVersionReviews",
        "summary": "Get version review statistics",
        "description": "View all reviews, reputation scores, and approval/rejection progress for a specific version",
        "parameters": [
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "description": "Version ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review statistics retrieved"
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/moderation/queue/{id}/approve": {
      "post": {
        "operationId": "ModerationController_approveFromQueue",
        "summary": "[ADMIN] Force approve item",
        "description": "Admin override: Directly approve a pending version bypassing community review process",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Moderation queue item ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveVersionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item approved successfully"
          },
          "403": {
            "description": "Admin privileges required"
          },
          "404": {
            "description": "Queue item not found"
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/moderation/queue/{id}/reject": {
      "post": {
        "operationId": "ModerationController_rejectFromQueue",
        "summary": "[ADMIN] Force reject item",
        "description": "Admin override: Directly reject a pending version bypassing community review process",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Moderation queue item ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectVersionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item rejected successfully"
          },
          "403": {
            "description": "Admin privileges required"
          },
          "404": {
            "description": "Queue item not found"
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/moderation/queue/{id}/assign": {
      "post": {
        "operationId": "ModerationController_assignModerator",
        "summary": "Assign moderator to queue item",
        "description": "Assign a moderator to review a specific queue item",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Moderation queue item ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Moderator assigned successfully"
          },
          "404": {
            "description": "Queue item not found"
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/moderation/queue/{id}": {
      "get": {
        "operationId": "ModerationController_getQueueItem",
        "summary": "Get queue item details",
        "description": "Retrieve detailed information about a specific moderation queue item",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Moderation queue item ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/moderation/stats": {
      "get": {
        "operationId": "ModerationController_getStats",
        "summary": "Get moderation statistics",
        "description": "Retrieve comprehensive moderation statistics and metrics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully"
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/moderation/history": {
      "get": {
        "operationId": "ModerationController_getHistory",
        "summary": "Get moderation history",
        "description": "Retrieve history of reviewed items",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "History retrieved successfully"
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/moderation/bulk/approve": {
      "post": {
        "operationId": "ModerationController_bulkApprove",
        "summary": "Bulk approve queue items",
        "description": "Approve multiple queue items at once",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Bulk operation completed"
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/moderation/bulk/reject": {
      "post": {
        "operationId": "ModerationController_bulkReject",
        "summary": "Bulk reject queue items",
        "description": "Reject multiple queue items at once",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Bulk operation completed"
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/moderation/top-moderators": {
      "get": {
        "operationId": "ModerationController_getTopModerators",
        "summary": "Get top moderators by activity",
        "description": "Retrieve list of most active moderators",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of top moderators to return",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/moderation/config": {
      "get": {
        "operationId": "ModerationController_getConfig",
        "summary": "Get moderation configuration",
        "description": "Retrieve current moderation thresholds and settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Configuration retrieved successfully"
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "ModerationController_updateConfig",
        "summary": "Update moderation configuration",
        "description": "Update moderation thresholds and settings. Only admins can modify these settings.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModerationConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuration updated successfully"
          },
          "403": {
            "description": "Admin privileges required"
          }
        },
        "tags": [
          "Collaboration - Moderation"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "♻️ Recycling Centers Management API",
    "description": "Enterprise-grade API for managing recycling centers, materials, and user operations",
    "version": "1.0",
    "contact": {
      "name": "The Wealthiest Programmer",
      "url": "https://github.com/wealthiest-programmer",
      "email": "contact@wealthiest-programmer.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateVersionReviewDto": {
        "type": "object",
        "properties": {}
      },
      "ApproveVersionDto": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string",
            "description": "Moderator notes for the approval",
            "example": "Approved after verification of coordinates"
          }
        }
      },
      "RejectVersionDto": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string",
            "description": "Moderator notes explaining the rejection",
            "example": "Coordinates appear to be incorrect"
          }
        }
      },
      "UpdateModerationConfigDto": {
        "type": "object",
        "properties": {
          "approvalThreshold": {
            "type": "number",
            "description": "Reputation threshold for auto-approval",
            "example": 100
          },
          "rejectionThreshold": {
            "type": "number",
            "description": "Reputation threshold for auto-rejection",
            "example": 100
          },
          "minReviewsRequired": {
            "type": "number",
            "description": "Minimum reviews required before auto-decision",
            "example": 1
          }
        }
      }
    }
  }
}